# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a server for Kuro API",
        "title": "Bookify",
        "contact": {
            "name": "API Support",
            "email": "hoaiphong01012002@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/activity-logs/get/id": {
            "get": {
                "description": "Retrieve details of an activity using its id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activity Logs"
                ],
                "summary": "Get Activity by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Activity Log ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/activity-logs/get/level": {
            "get": {
                "description": "Retrieve details of an organization using its Level.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activity Logs"
                ],
                "summary": "Get organization by level",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Activity Level",
                        "name": "level",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
...
}